requirements.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-tables2==1.3.0
pkg-resources==0.0.0
